+++ src/main/java/com/example/librarydemo/DTO/StudentsDTO.java
@@ -1,14 +1,14 @@
 package com.example.librarydemo.DTO;
 
 public interface StudentsDTO {
-    int getId();
+    long getId();
     String getEmail();
     boolean getEnabled();
     String getFull_name();
     String getLogin();
     String getPassword();
     String getPhone_number();
-    int getRole_id();
+    long getRole_id();
     String getAddress();
     String getGrop();
 }
