+++ src/main/java/com/example/librarydemo/repository/EBookRepository.java
@@ -4,7 +4,7 @@ import com.example.librarydemo.models.EBook;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EBookRepository extends CrudRepository<EBook, Integer> {
+public interface EBookRepository extends CrudRepository<EBook, Long> {
     @Query(value="SELECT COUNT(id) FROM `e_book`", nativeQuery = true)
     Integer getEBookQuantity();
 }
